refactor code + update readme
