+++ README.md
@@ -1,30 +1,29 @@
-# React + TypeScript + Vite
+# ReactScreen
+## React + TypeScript + Vite
 
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
+<p>В данном проекте представлен экран, реализованный с помощью React + TypeScript + Vite.<p>
 
-Currently, two official plugins are available:
+### Скриншоты
 
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
+<table>
+    <tr>
+        <td><img src="screens/image.png"></td>
+    </tr>
+    <tr>
+      <td><img src = "screens/image-2.png"></td>
+    </tr>
+    <tr>
+      <td><img src = "screens/image-3.png"></td>
+    </tr>
+    <tr>
+      <td><img src = "screens/image-4.png"></td>
+    </tr>
+</table>
 
-## Expanding the ESLint configuration
+### Технологии
 
-If you are developing a production application, we recommend updating the configuration to enable type aware lint rules:
-
-- Configure the top-level `parserOptions` property like this:
-
-```js
-export default {
-  // other rules...
-  parserOptions: {
-    ecmaVersion: 'latest',
-    sourceType: 'module',
-    project: ['./tsconfig.json', './tsconfig.node.json'],
-    tsconfigRootDir: __dirname,
-  },
-}
-```
-
-- Replace `plugin:@typescript-eslint/recommended` to `plugin:@typescript-eslint/recommended-type-checked` or `plugin:@typescript-eslint/strict-type-checked`
-- Optionally add `plugin:@typescript-eslint/stylistic-type-checked`
-- Install [eslint-plugin-react](https://github.com/jsx-eslint/eslint-plugin-react) and add `plugin:react/recommended` & `plugin:react/jsx-runtime` to the `extends` list
+- npm 10.2.3
+- React 18.2.0
+- TypeScript 5.2.2
+- Vite 5.0.8
+- Приложение разрабатывалось в Visual Studio Code
