+++ src/app/css/components/Articles.css
@@ -4,24 +4,24 @@
     width: inherit;
     align-items: center;
     justify-content: center;
-  }
+}
   
-  .title-articles{
+.title-articles{
   font-weight: 700;
   color: #000000;
   font-size: 36px;
   letter-spacing: 0;
   line-height: 56px;
   white-space: nowrap;
-  }
+}
   
-  .array-articles{
+.array-articles{
     display: flex;
     flex-wrap: wrap;
     margin-left: 240px;
-  }
+}
   
-  .article{
+.article{
     margin-right: 34px;
     margin-bottom: 37px;
     width: 350px;
@@ -29,16 +29,17 @@
     background-color: #ffffff;
     border-radius: 20px;
     box-shadow: 10px 40px 50px 0px rgba(229, 233, 246, 0.4);
-  }
+}
   
-  .article-image{
+.article-image{
     border-top-right-radius: 20px;
     border-top-left-radius: 20px;
     width: 350px;
     height: 240px;
     object-fit: cover;
-  }
-  .article-title{
+}
+
+.article-title{
     margin-top: 21px;
     margin-bottom: 12px;
     margin-left: 33px;
@@ -48,9 +49,9 @@
     font-size: 21px;
     letter-spacing: 0;
     line-height: 32px;
-  
-  }
-  .article-desc{
+}
+
+.article-desc{
     width: 266px;
     font-weight: 300;
     color: #7c7987;
@@ -60,30 +61,30 @@
     margin-left: 33px;
     margin-right: 66px;
     margin-bottom: 25px;
-  }
-  .read-more{
+}
+
+.read-more{
     margin-left: 33px;
-  }
-  .read-more-text{
+}
+
+.read-more-text{
     font-weight: 600;
     color: #3f88ed;
     font-size: 17px;
     letter-spacing: 0;
     line-height: 28px;
     text-decoration: none;
-
-  }
+}
   
-  .wrap-btn-view-all{
+.wrap-btn-view-all{
     width: 85%;
     display: flex;
     align-items: center;
     justify-content: center;
     margin-top: 33px;
+}
   
-  }
-  
-  .btn-view-all{
+.btn-view-all{
     width: 200px;
     height: 56px;
     border-radius: 55px;
@@ -93,13 +94,12 @@
     color: #458ff6;
     font-weight: 700;
     font-size: 18px;
-  }
+}
   
-  .article-arrow{
+.article-arrow{
     margin-left: 8px;
-  }
+}
 
-  div a:hover{
+div a:hover{
     color: #b7baf8;
-
-  }
+}
