+++ src/app/css/components/Footer.css
@@ -6,18 +6,18 @@
     height: 430px;
     background: linear-gradient(180deg, rgb(103, 195, 243) 0%, rgb(90, 152, 242) 100%);
     overflow:hidden;
-  }
+}
   
-  .ellipse-footer {
+.ellipse-footer {
     margin-left: 191px;
     margin-top: 118px;
     width: 41px;
     height: 41px;
     background-color: #ffffff;
     border-radius: 20.5px;
-    }
+}
   
-  .logo-letter-footer{
+.logo-letter-footer{
     display: block;
     margin-left: 13px;
     margin-top: 4px;
@@ -26,8 +26,9 @@
     color: #5b9bf3;
     font-size: 26px;
     letter-spacing: 0;
-    }
-  .title-footer{
+}
+
+.title-footer{
     margin-top: 124px;
     margin-left: 12px;
     font-weight: 700;
@@ -35,9 +36,9 @@
     font-size: 24px;
     letter-spacing: 0;
     line-height: normal;
-  }
+}
   
-  .text-about-company{
+.text-about-company{
     margin-left: 190px;
     width: 391px;
     font-weight: 300;
@@ -45,9 +46,9 @@
     font-size: 18px;
     letter-spacing: 0;
     line-height: 28px;
-  }
+}
   
-  .footer-company{
+.footer-company{
     margin-top: 31px;
     margin-left: 190px;
     width: 391px;
@@ -56,9 +57,9 @@
     font-size: 16px;
     letter-spacing: 0;
     line-height: 28px;
-  }
+}
   
-  .catalog-footer{
+.catalog-footer{
     margin-top: 118px;
     margin-left: 145px;
     font-weight: 700;
@@ -67,22 +68,22 @@
     letter-spacing: 0;
     line-height: 50px;
     white-space: nowrap;
-  }
+}
   
-  .list-links{
+.list-links{
     margin-left: 105px;
     list-style-type:none;
-  }
+}
   
-  .list-links li a{
+.list-links li a{
     text-decoration: none;
     font-weight: 300;
-  color: #ffffff;
-  font-size: 18px;
-  letter-spacing: 0;
-  line-height: 38px;
-  }
+    color: #ffffff;
+    font-size: 18px;
+    letter-spacing: 0;
+    line-height: 38px;
+}
   
-  .wrap-elem-footer{
+.wrap-elem-footer{
     margin-bottom: 50px;
-  }
+}
