+++ src/app/css/components/Header.css
@@ -1,17 +1,17 @@
 .head{
     display: flex;
-  }
+}
   
-  .ellipse {
+.ellipse {
     margin-left: 191px;
     margin-top: 51px;
     width: 41px;
     height: 41px;
     background-color: #458ff6;
     border-radius: 20.5px;
-    }
+}
   
-  .logo-letter{
+.logo-letter{
     display: block;
     margin-left: 13px;
     margin-top: 4px;
@@ -20,9 +20,9 @@
     color: #ffffff;
     font-size: 26px;
     letter-spacing: 0;
-    }
+}
   
-  .title{
+.title{
     margin-top: 55px;
     margin-left: 12px;
     font-weight: 700;
@@ -32,6 +32,6 @@
     line-height: normal;
     }
   
-  .ellipse:hover {
+.ellipse:hover {
     filter: drop-shadow(0 0 2em #646cffaa);
-  }
+}
