+++ src/app/css/components/Providers.css
@@ -3,39 +3,39 @@
     margin-top: 220px;
     margin-left: 135px;
     margin-right: 192px;
-  }
+}
   
-  .left-side-provider{
+.left-side-provider{
     display: block;
     margin-right: 142px;
   
-  }
+}
   
-  .wrap-divider-provider{
+.wrap-divider-provider{
     width:inherit;
     margin-top: 26px;
     margin-bottom: 33px;
-  }
+}
   
-  .provider-header{
+.provider-header{
     width: 481px;
     font-weight: 700;
     color: #000000;
     font-size: 36px;
     letter-spacing: 0;
     line-height: 56px;
-  }
+}
   
-  .text-provider{
+.text-provider{
     width: 450px;
     font-weight: 300;
     color: #7c7987;
     font-size: 18px;
     letter-spacing: 0;
     line-height: 30px;
-  }
+}
   
-  .wrap-btn-provider{
+.wrap-btn-provider{
     width:inherit;
     display: flex;
     align-items: start;
@@ -43,12 +43,12 @@
     margin-top: 26px;
     margin-bottom: 33px;
   
-  }
+}
   
-  .wrap-elem-provider{
+.wrap-elem-provider{
     display: flex;
     width:inherit;
     align-items:end;
     justify-content: end;
   
-  }
+}
