+++ src/app/css/components/Services.css
@@ -6,25 +6,25 @@
     letter-spacing: 0;
     line-height: 56px;
     white-space: nowrap;
-  }
+}
   
-  .wrap-divider-services{
+.wrap-divider-services{
     width:inherit;
     display: flex;
     align-items: center;
     justify-content: center;
     margin-top: 26px;
     margin-bottom: 33px;
-  }
+}
   
-  .divider{
+.divider{
     width: 56px;
     height: 2px;
     background-color: #000000;
     border-radius: 5px;
-  }
+}
   
-  .disc-services{
+.disc-services{
     margin-left: 250px;
     margin-right: 270px;
     font-weight: 300;
@@ -33,66 +33,66 @@
     text-align: center;
     letter-spacing: 0;
     line-height: 30px;
-  }
+}
   
-  .array-cards{
+.array-cards{
     display:flex;
     flex-wrap: wrap;
     margin-left: 240px;
     margin-top: 80px;
     margin-bottom: -90px;
-  }
+}
   
-  .background-class{
+.background-class{
     z-index: -2;
     background: url('/src/assets/images/background.svg');
     background-size: 764px 619px;
     background-repeat: no-repeat;
-  }
+}
   
-  .element-background{
+.element-background{
     z-index: -2;
     position: relative;
     left: 1035px;
     bottom: 495px;
-  }
+}
   
-  .card{
-  width: 350px;
-  height: 354px;
-  background-color: #ffffff;
-  border-radius: 20px;
-  margin-bottom: 37px;
-  margin-right: 34px;
-  box-shadow: 10px 40px 50px 0px rgba(229, 233, 246, 0.4);
-  }
+.card{
+    width: 350px;
+    height: 354px;
+    background-color: #ffffff;
+    border-radius: 20px;
+    margin-bottom: 37px;
+    margin-right: 34px;
+    box-shadow: 10px 40px 50px 0px rgba(229, 233, 246, 0.4);
+}
   
-  .card-image{
+.card-image{
     margin-top: 50px;
     margin-left: 37px;
-  }
+}
   
-  .card-title{
-  font-weight: 700;
-  color: #000000;
-  font-size: 24px;
-  letter-spacing: 0;
-  line-height: 56px;
-  margin-left: 39px;
+.card-title{
+    font-weight: 700;
+    color: #000000;
+    font-size: 24px;
+    letter-spacing: 0;
+    line-height: 56px;
+    margin-left: 39px;
   
-  }
+}
   
-  .card-desc{
-  width: 266px;
-  font-weight: 300;
-  color: #7c7987;
-  font-size: 16px;
-  letter-spacing: 0;
-  line-height: 28px;
-  margin-left: 39px;
-  }
+.card-desc{
+    width: 266px;
+    font-weight: 300;
+    color: #7c7987;
+    font-size: 16px;
+    letter-spacing: 0;
+    line-height: 28px;
+    margin-left: 39px;
+}
   
-  .btn-learnmore{
+.btn-learnmore{
     width: 200px;
     height: 56px;
     border-radius: 55px;
@@ -103,10 +103,10 @@
     font-weight: 700;
     font-size: 18px;
     white-space: nowrap;
-  }
-  .wrap-btn-services{
+}
+.wrap-btn-services{
     width:inherit;
     display: flex;
     align-items: center;
     justify-content: center;
-  }
+}
