+++ src/app/css/components/Slider.css
@@ -9,7 +9,7 @@
     overflow: hidden;
   }
   
-  .slider-card-title{
+.slider-card-title{
     font-weight: 700;
     color: #ffffff;
     font-size: 36px;
@@ -17,39 +17,40 @@
     line-height: 48px;
     white-space: nowrap;
     padding-left: 300px;
-    padding-top: 64px;
-    
-  }
+    padding-top: 64px; 
+}
   
   
-  .elem-slider{
+.elem-slider{
     position:absolute;
     right: 168px;
-  
-  }
-  .wrap-divider-slider{
+}
+
+.wrap-divider-slider{
     margin-top: 25px;
     display: flex;
     width:inherit;
     align-items:center;
     justify-content: center;
-  
-  }
-  .divider-white{
+}
+
+.divider-white{
     width: 56px;
     height: 2px;
     border-radius: 5px;
     background-color: white;
-  }
-  .author{
+}
+
+.author{
     display: flex;
     margin-left: 110px;
-  }
-  .comment-block{
+}
+
+.comment-block{
     margin-top: 65px;
-  }
+}
   
-  .wrap-avatar{
+.wrap-avatar{
     width: 141px;
     height: 141px;
     overflow:hidden;
@@ -57,10 +58,9 @@
     border-radius: 50%;
     border: 5px solid white;
     margin-right: 29px;
+}
   
-  }
-  
-  .avatar{
+.avatar{
     position: relative;
     right:90px;
     bottom: 120px;
@@ -68,41 +68,40 @@
     height: auto;
     object-fit:cover;
     border-radius: 50%;
-  }
+}
   
-  .name-author{
+.name-author{
     margin-top: 35px;
     font-weight: 700;
     color: #ffffff;
     font-size: 22px;
     letter-spacing: 0;
     white-space: nowrap;
-  }
+}
   
-  .job-author{
+.job-author{
     margin-top: 10px;
     font-weight: 400;
     color: #ffffffd9;
     font-size: 18px;
     letter-spacing: 0;
     white-space: nowrap;
-  }
+}
   
-  .text-slider{
+.text-slider{
+    font-weight: 400;
+    color: #ffffffe6;
+    font-size: 19px;
+    letter-spacing: 0;
+    line-height: 30px;
+}
   
-  font-weight: 400;
-  color: #ffffffe6;
-  font-size: 19px;
-  letter-spacing: 0;
-  line-height: 30px;
-  }
-  
-  .wrap-text-slider{
+.wrap-text-slider{
     width: 420px;
     margin-left: 110px;
-  }
+}
   
-  .button-left{
+.button-left{
     background-image: url('/src/assets/images/arrow-right.svg');
     background-size: cover;
     transform: scale(-1, 1);
@@ -111,9 +110,9 @@
     margin-right: 60px;
     border: none;
     background-color: #ffffff;
-  }
+}
   
-  .button-right{
+.button-right{
     border: none;
     background-image: url('/src/assets/images/arrow-right.svg');
     background-size: cover;
@@ -121,27 +120,27 @@
     height: 18px;
     margin-left: 90px;
     background-color: #ffffff;
-  }
-  .button-right.disabled{
+}
+
+.button-right.disabled{
     background-image: url('/src/assets/images/arrow-left.svg');
     transform: scale(-1, 1);
+}
   
-  }
-  
-  .button-left.disabled{
+.button-left.disabled{
     background-image: url('/src/assets/images/arrow-left.svg');
     transform: scale(1, -1);
-  }
+}
   
-  .wrap-manage-slider{
+.wrap-manage-slider{
     width: inherit;
     align-items: center;
     justify-content: center;
     display: flex;
     margin-top: 40px;
-  }
+}
   
-  .circle {
+.circle {
     width: 10px;
     height: 10px;
     margin: 0 3px;
@@ -150,12 +149,12 @@
     border-radius: 50%;
     cursor: pointer;
     margin-left: 30px;
-  }
+}
   
-  .circle.selected {
+.circle.selected {
     background-color: #3498db;
-  }
+}
   
-  .about-author{
+.about-author{
     width: 170px;
-  }
+}
