+++ src/app/css/components/Subheader.css
@@ -1,41 +1,43 @@
 .sub-header{
     display: flex;
-  }
-  .left-side-virtual-healthcare{
+}
+
+.left-side-virtual-healthcare{
     display: block;
     margin-left: 191px;
     margin-right: 127px;
-  }
-  .sub-header-illustration {
+}
+
+.sub-header-illustration {
     width: 693px;
     height: 598px;
     object-fit: cover;
-    }
+}
   
-  .virtual-healthcare {
+.virtual-healthcare {
     font-weight: 700;
     color: #000000;
     font-size: 48px;
     letter-spacing: 0;
     line-height: 56px;
-  }
+}
   
-  .text-sub-header {
+.text-sub-header {
     width: 427px;
     font-weight: 300;
     color: #7c7987;
     font-size: 21px;
     letter-spacing: 0;
     line-height: 32px;
-  }
+}
   
-  .right-side-virtual-healthcare{
+.right-side-virtual-healthcare{
     position: relative;
     bottom: 100px;
     right: 30px;
-  }
+}
   
-  .consult-btn {
+.consult-btn {
     width: 200px;
     height: 56px;
     background-color: #458ff6;
@@ -46,13 +48,13 @@
     letter-spacing: 0;
     white-space: nowrap;
     border-width: 0;
-  }
+}
   
-  .consult-btn:hover {
+.consult-btn:hover {
     border-color: #646cff;
-  }
+}
   
-  .element{
+.element{
     margin-top: 57px;
     margin-bottom: 8px;
-  }
+}
